Project structure refactoring. Moved cypress and required resource for tests into test/ folder.
