+++ package.json
@@ -4,11 +4,11 @@
   "description": "Cypress commands for Cumulocity IoT",
   "main": "lib/commands/index.js",
   "scripts": {
-    "start-server": "http-server app/ --port 8080 --silent --mimetypes custom_mime_types.types",
+    "start-server": "http-server test/cypress/app/ --port 8080 --silent --mimetypes test/custom_mime_types.types",
     "test": "TZ=Europe/Berlin start-server-and-test start-server http://localhost:8080 test:run",
-    "test:run": "TZ=Europe/Berlin cypress run --browser chrome --config video=false",
+    "test:run": "TZ=Europe/Berlin cypress run  --project test --browser chrome --config video=false",
     "test:open": "TZ=Europe/Berlin start-server-and-test start-server http://localhost:8080 debug",
-    "debug": "CYPRESS_REMOTE_DEBUGGING_PORT=9222 cypress open",
+    "debug": "CYPRESS_REMOTE_DEBUGGING_PORT=9222 cypress open --project test",
     "clean": "rimraf dist/ && rimraf runner/dist/",
     "copy-files": "copyfiles lib/**/*.*js lib/**/*.d.ts README.md package.json dist/",
     "build": "npm run clean && npm run copy-files && tsc && cd dist/ && npm run clean-package-json",
