+++ test/cypress.config.js
@@ -1,5 +1,5 @@
 const { defineConfig } = require("cypress");
-const registerC8yPlugin = require("./lib/plugin");
+const registerC8yPlugin = require("../lib/plugin");
 
 module.exports = defineConfig({
   e2e: {
