+++ test/cypress/fixtures/c8ypact-managedobject-01.json
