+++ test/cypress/support/e2e.ts
@@ -14,5 +14,5 @@
 // ***********************************************************
 
 // Import commands.js using ES2015 syntax:
-import "../../lib/commands"
-import "../../lib/commands/request"
+import "../../../lib/commands"
+import "../../../lib/commands/request"
